refactor(storage): extract AgentRunStore sub-mixin (SQLite + ABC)
Task 2 of the _agent_run.py decomposition — the first extraction, which
establishes the agent_run/ package + composite-ABC + leaf-_models +
residual-holder pattern for the remaining PTC/RTD tasks.

Move the AgentRunStore bucket (7 lifecycle methods + the 3 run-cascade
_..._unlocked privates) out of the SQLite/ABC AgentRun god-class into a new
cohesive sub-mixin, VERBATIM (byte-identical method bodies; only the
@handle_exceptions decorator + import lines are rewritten):

- storage_base/agent_run/_models.py  — leaf holding the 18 shared non-class
  symbols (8 models/enums + 8 helpers + NOT_APPLICABLE_ANSWER + the
  PendingToolCallStatus re-export); imports no mixin (no cycle).
- storage_base/agent_run/_agent_run_store.py — AgentRunStoreABC (7 stubs).
- storage_base/_agent_run.py — AgentRunMixin survives as a thin COMPOSITE
  (AgentRunMixin(AgentRunStoreABC) + the remaining PTC/RTD stubs + the public
  helper staticmethods) and re-exports the 18 models, so
  storage_base/__init__.py + BaseStorage(AgentRunMixin, ...) stay byte-identical.
- sqlite_storage/agent_run/_agent_run_store.py — SQLiteAgentRunStoreMixin
  (@SQLiteStorageBase.handle_exceptions), composed AHEAD of the shrinking
  residual SQLiteAgentRunMixin. The 3 privates stay SINGLE here and are reached
  cross-bucket by the residual pending-tool-call methods via MRO co-composition
  (pyright-only Callable stubs on the consumer).

The shared row/datetime helpers (_dt, _dt_str, _row_to_agent_run, …) stay in
the residual modules — the PTC/RTD methods still use them — so no test import
repoints are needed on the SQLite side.
