+++ reflexio/server/services/storage/sqlite_storage/__init__.py
@@ -24,6 +24,7 @@ from ._stall_state import (
     mark_stall_notified,
     upsert_stall_state,
 )
+from .agent_run import SQLiteAgentRunStoreMixin
 from .base import SQLiteDeletionMixin, SQLiteFtsVecMixin
 from .governance import (
     AuditEventStoreMixin,
@@ -43,6 +44,7 @@ from .profiles import InteractionStoreMixin, ProfileSearchMixin, ProfileStoreMix
 
 
 class SQLiteStorage(
+    SQLiteAgentRunStoreMixin,
     SQLiteAgentRunMixin,
     ProfileStoreMixin,
     InteractionStoreMixin,
